WIP: Factor region rendering to its own class
